Fix IMU reading of signed acceleration values

+++ vehicle_interface/src/sd_vehicle_interface/sd_gps_imu.cpp
@@ -142,9 +142,9 @@ namespace sd{
 		switch (ReceivedFrameCAN.id){
 			case 1536: //0x600
 				{				
-					IMU_Accel_X = ReceivedFrameUnion.word[0]*3.91*0.001*9.80665; // 0.038344002; //*3.91 = mG, then conversion to m/s^2
-					IMU_Accel_Y = ReceivedFrameUnion.word[1]*3.91*0.001*9.80665; //
-					IMU_Accel_Z = ReceivedFrameUnion.word[2]*3.91*0.001*9.80665; //
+					IMU_Accel_X = ((int16_t)ReceivedFrameUnion.word[0])*3.91*0.001*9.80665; // 0.038344002; //*3.91 = mG, then conversion to m/s^2
+					IMU_Accel_Y = ((int16_t)ReceivedFrameUnion.word[1])*3.91*0.001*9.80665; //
+					IMU_Accel_Z = ((int16_t)ReceivedFrameUnion.word[2])*3.91*0.001*9.80665; //
 
 					break;
 				}
